[UPDATE] 예약페이지 운임 정보, 예약 vo, controller, [FIX] 검색페이지 날짜정보

+++ src/main/java/com/nadri/train/vo/TrainReservation.java
@@ -13,7 +13,7 @@ public class TrainReservation {
 	
 	private int no;
 	private int totalCount;
-	private int totalPrice;
+	private long totalPrice;
 	private String tickectStatus;
 	private Date canceledDate;
 	private Date soldDate;
